remove mod high_level_rs/low_level_rs
move high_level_rs members to api
move low_level_rs::spring_config mod to api
remove fn spring_config_toml
