--- springql-core/src/api/low_level_rs.rs
@@ -1,9 +1,0 @@
-// This file is part of https://github.com/SpringQL/SpringQL which is licensed under MIT OR Apache-2.0. See file LICENSE-MIT or LICENSE-APACHE for full license details.
-
-//! Low-level API functions to execute / register SpringQL from Rust.
-//!
-//! C API and high-level Rust API are provided separately.
-
-pub mod spring_config;
-
-pub use crate::api::low_level_rs::spring_config::*;
