--- springql-core/src/api/low_level_rs/engine_mutex.rs
@@ -1,1 +1,0 @@
-// This file is part of https://github.com/SpringQL/SpringQL which is licensed under MIT OR Apache-2.0. See file LICENSE-MIT or LICENSE-APACHE for full license details.
